+++ core/src/main/java/io/lionweb/lioncore/java/serialization/FlatBuffersSerialization.java
@@ -1,10 +1,8 @@
 package io.lionweb.lioncore.java.serialization;
 
 import com.google.flatbuffers.FlatBufferBuilder;
-import io.lionweb.lioncore.java.language.LionCoreBuiltins;
 import io.lionweb.lioncore.java.model.ClassifierInstance;
 import io.lionweb.lioncore.java.model.impl.ProxyNode;
-import io.lionweb.lioncore.java.self.LionCore;
 import io.lionweb.lioncore.java.serialization.data.*;
 import io.lionweb.lioncore.java.serialization.data.MetaPointer;
 import io.lionweb.lioncore.java.serialization.flatbuffers.*;
@@ -16,24 +14,6 @@ import java.util.stream.Collectors;
 public class FlatBuffersSerialization extends AbstractSerialization {
 
   private static final String NULL_CONSTANT = "NULL";
-
-  /** This has specific support for LionCore or LionCoreBuiltins. */
-  public static FlatBuffersSerialization getStandardSerialization() {
-    FlatBuffersSerialization serialization = new FlatBuffersSerialization();
-    serialization.classifierResolver.registerLanguage(LionCore.getInstance());
-    serialization.instantiator.registerLionCoreCustomDeserializers();
-    serialization.primitiveValuesSerialization
-        .registerLionBuiltinsPrimitiveSerializersAndDeserializers();
-    serialization.instanceResolver.addAll(LionCore.getInstance().thisAndAllDescendants());
-    serialization.instanceResolver.addAll(LionCoreBuiltins.getInstance().thisAndAllDescendants());
-    return serialization;
-  }
-
-  /** This has no specific support for LionCore or LionCoreBuiltins. */
-  public static FlatBuffersSerialization getBasicSerialization() {
-    FlatBuffersSerialization serialization = new FlatBuffersSerialization();
-    return serialization;
-  }
 
   public List<io.lionweb.lioncore.java.model.Node> deserializeToNodes(byte[] bytes)
       throws IOException {