+++ core/src/test/java/io/lionweb/lioncore/java/serialization/ProtobufSerializationTest.java
@@ -61,7 +61,7 @@ public class ProtobufSerializationTest extends SerializationTest {
   public void deserializeMultipleRoots() throws IOException {
     Sum sum1 = new Sum(new IntLiteral(1), new IntLiteral(2));
     Sum sum2 = new Sum(new IntLiteral(3), new IntLiteral(4));
-    ProtoBufSerialization serialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization serialization = SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = serialization.serializeTreesToByteArray(sum1, sum2);
     prepareDeserializationOfSimpleMath(serialization);
     List<Sum> deserialized =
@@ -78,7 +78,7 @@ public class ProtobufSerializationTest extends SerializationTest {
     IntLiteral il2 = new IntLiteral(2, null);
     IntLiteral il3 = new IntLiteral(3, null);
     IntLiteral il4 = new IntLiteral(4, null);
-    ProtoBufSerialization serialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization serialization = SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = serialization.serializeTreesToByteArray(il1, il2, il3, il4);
     prepareDeserializationOfSimpleMath(serialization);
     List<IntLiteral> deserialized =
@@ -96,7 +96,7 @@ public class ProtobufSerializationTest extends SerializationTest {
     IntLiteral il3 = new IntLiteral(3, "int_3");
     IntLiteral il4 = new IntLiteral(4, "int_4");
     Sum sum2 = new Sum(il3, il4, null);
-    ProtoBufSerialization serialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization serialization = SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = serialization.serializeTreesToByteArray(sum1, sum2);
     prepareDeserializationOfSimpleMath(serialization);
     List<Node> deserialized = serialization.deserializeToNodes(serialized);
@@ -113,7 +113,7 @@ public class ProtobufSerializationTest extends SerializationTest {
     IntLiteral il3 = new IntLiteral(3, "int_3");
     IntLiteral il4 = new IntLiteral(4, "int_4");
     Sum sum2 = new Sum(il3, il4, null);
-    ProtoBufSerialization serialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization serialization = SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = serialization.serializeNodesToByteArray(il4, il1, sum1, il2, sum2, il3);
     prepareDeserializationOfSimpleMath(serialization);
     List<Node> deserialized = serialization.deserializeToNodes(serialized);
@@ -126,7 +126,7 @@ public class ProtobufSerializationTest extends SerializationTest {
     IntLiteral il1 = new IntLiteral(1, "int_1");
     IntLiteral il2 = new IntLiteral(2, null);
     Sum sum1 = new Sum(il1, il2, null);
-    ProtoBufSerialization serialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization serialization = SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = serialization.serializeNodesToByteArray(sum1, il1, il2);
     prepareDeserializationOfSimpleMath(serialization);
     List<Node> deserialized = serialization.deserializeToNodes(serialized);
@@ -157,7 +157,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     RefNode r1 = new RefNode();
     RefNode r2 = new RefNode();
     r1.setReferred(r2);
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = protoBufSerialization.serializeNodesToByteArray(r1);
     prepareDeserializationOfRefMM(protoBufSerialization);
     List<Node> deserialized = protoBufSerialization.deserializeToNodes(serialized);
@@ -171,7 +172,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     r1.setReferred(r2);
     r2.setReferred(r3);
     r3.setReferred(r1);
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = protoBufSerialization.serializeNodesToByteArray(r1, r2, r3);
     prepareDeserializationOfRefMM(protoBufSerialization);
     List<Node> deserialized = protoBufSerialization.deserializeToNodes(serialized);
@@ -192,7 +194,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     Assert.assertEquals(Arrays.asList(c2), ClassifierInstanceUtils.getChildren(c1));
     Assert.assertEquals(Arrays.asList(c1), ClassifierInstanceUtils.getChildren(c2));
 
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     byte[] serialized = protoBufSerialization.serializeNodesToByteArray(c1, c2);
     prepareDeserializationOfRefMM(protoBufSerialization);
     List<Node> deserialized = protoBufSerialization.deserializeToNodes(serialized);
@@ -208,7 +211,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     myLanguage.addElement(myConcept);
 
     DynamicNode myInstance = new DynamicNode("instance-a", myConcept);
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     SerializedChunk serializedChunk =
         protoBufSerialization.serializeNodesToSerializationBlock(myInstance);
     assertEquals(1, serializedChunk.getClassifierInstances().size());
@@ -234,7 +238,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     AnnotationInstance a1_2 = new DynamicAnnotationInstance("a1_2", a1, n1);
     AnnotationInstance a2_3 = new DynamicAnnotationInstance("a2_3", a2, n1);
 
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     protoBufSerialization.enableDynamicNodes();
     SerializedChunk serializedChunk = protoBufSerialization.serializeNodesToSerializationBlock(n1);
 
@@ -273,7 +278,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     DynamicAnnotationInstance ann = new DynamicAnnotationInstance("metaAnn_1", metaAnn, c);
     c.addAnnotation(ann);
 
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     protoBufSerialization.enableDynamicNodes();
     SerializedChunk serializedChunk = protoBufSerialization.serializeTreeToSerializationBlock(l);
 
@@ -308,7 +314,8 @@ public class ProtobufSerializationTest extends SerializationTest {
     DynamicNode n1 = new DynamicNode("n1", c);
     ClassifierInstanceUtils.setPropertyValueByName(n1, "foo", "abc");
 
-    ProtoBufSerialization protoBufSerialization = ProtoBufSerialization.getStandardSerialization();
+    ProtoBufSerialization protoBufSerialization =
+        SerializationProvider.getStandardProtoBufSerialization();
     SerializedChunk serializedChunk = protoBufSerialization.serializeNodesToSerializationBlock(n1);
 
     assertEquals(2, serializedChunk.getLanguages().size());